hoàn thiện admin

+++ Admin/WEBBANSACH/WEBBANSACH/Models/Sach.cs
@@ -9,12 +9,9 @@ namespace WEBBANSACH.Models
         public int TacGiaID { get; set; } // Khóa ngoại tham chiếu đến TacGias
         public int NhaXuatBanID { get; set; } // Khóa ngoại tham chiếu đến NhaXuatBans
         public int TheLoaiID { get; set; } // Khóa ngoại tham chiếu đến TheLoais
-
-        // Chuyển kiểu dữ liệu từ float sang decimal để tương thích với DECIMAL trong SQL
-        public decimal Sao { get; set; } // Số sao
-        public decimal Sale { get; set; } // Giảm giá
-        public decimal Gia { get; set; } // Giá
-
+        public Double Sao { get; set; } // Số sao
+        public Double Sale { get; set; } // Giảm giá
+        public Double Gia { get; set; } // Giá
         public int DiemThuong { get; set; } // Điểm thưởng
         public int DaBan { get; set; } // Số lượng đã bán
         public int SoLuong { get; set; } // Số lượng còn lại
